ssl-checker/requirements.txt
fastapi
uvicorn[standard]
requests
dnspython